modules/mini_app/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
openai